fix student features

+++ backend/src/main/java/com/tracker/studentracker/controllers/StudentController.java
@@ -26,7 +26,7 @@ public class StudentController {
 
         try{
 
-            int userId = userService.registerStudentUser(
+            Long userId = userService.registerStudentUser(
                     request.getName(),
                     request.getEmail(),
                     request.getPassword()
